specfied the default type for initalizing arrays in get_constants_positions() (appearently different default type on different versions and/or machines)
git-svn-id: https://svn.urbansim.org/trunk@14736 1513fa5d-83ce-422e-ae47-6967bb0fb594

+++ opus_core/specified_coefficients.py
@@ -411,7 +411,7 @@ class SpecifiedCoefficients(object):
             if var == self.constant_string:
                 result.append(i)
             i+=1
-        return array(result)
+        return array(result, dtype="int32")
 
     def add_calibration_constants(self, constants_array):
         const_index = self.get_constants_positions()[0]
